emulation: qemu-x86: Enable USB early
On x86 the keyboard is emulated as i8042 compatible controller.

When the keyboard is attached to an USB controller, for example by
specifying `-device usb-kbd,bus=xhci.0` on the command line, the i8042 is
still present, but doesn't do anything.

In order to use the U-Boot shell enable USB in board_late_init() as
done on other qemu targets.

TEST=Can enter the U-Boot shell when USB keyboard is attached to xHCI
     on qemu-pc target.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>